# docs/swagger.yaml
basePath: /
definitions:
...
  api.adminChallengesResponse:
    properties:
      challenges:
        items:
          $ref: '#/definitions/models.AdminChallengeView'
        type: array
      hasMore:
        description: |-
          HasMore says whether another page exists — what an infinite-scrolling
          client actually needs, so it never has to do the arithmetic.
        type: boolean
      limit:
        description: Limit is the window size actually applied; omitted when unpaginated.
        type: integer
      page:
        description: Page is 1-based; omitted when the caller asked for no page.
        type: integer
      total:
        description: Total is the row count across every page.
        type: integer
    type: object
...
  httpx.ErrorBody:
    properties:
      error:
        $ref: '#/definitions/httpx.ErrorDetail'
    type: object
  httpx.ErrorDetail:
    properties:
      code:
        type: string
      message:
        type: string
      request_id:
        type: string
    type: object
...
  models.AdminChallengeView:
    properties:
      category:
        type: string
      challengeDate:
        type: string
      challengeEventsEnabled:
        description: |-
          ChallengeEventsEnabled turns on the in-game disruption events (ink splash,
          fog, earthquake, quick countdown) for this challenge. Same value for every
          player attempting the day, so the shared leaderboard stays comparable.
        type: boolean
      createdAt:
        type: string
      gridCols:
        type: integer
      gridRows:
        type: integer
      id:
        type: string
      image:
        $ref: '#/definitions/models.Image'
      playSeconds:
        type: integer
      quizCorrectOption:
        type: integer
      quizOptionA:
        type: string
      quizOptionB:
        type: string
      quizQuestion:
        description: |-
          The bonus quiz asked after the puzzle is solved: one question, two
          options, one of them right. Optional — an empty question means the
          manager configured no quiz for this challenge.
        type: string
      title:
        type: string
      xpReward:
        type: integer
    type: object
...
  models.Image:
    properties:
      createdAt:
        type: string
      height:
        type: integer
      id:
        type: string
      isChallenge:
        type: boolean
      publicId:
        type: string
      thumbUrl:
        description: |-
          ThumbURL delivers the same picture resized and re-encoded for a list
          row or grid tile. Clients showing a small copy should read this instead
          of URL: it is typically a twentieth of the bytes.
        type: string
      title:
        type: string
      url:
        type: string
      width:
        type: integer
    type: object
...
info:
  contact: {}
  description: |-
    Backend for the Jigsaw Puzzle app: email/password auth, image
    collection, admin daily challenges (Cloudinary) and per-day
    reputation ("Danh vọng").
  title: Jigsaw Puzzle API
  version: "1.0"
paths:
  /api/v1/admin/challenges:
    get:
      parameters:
      - description: Page number (1-based); needs limit or defaults to 8
        in: query
        name: page
        type: integer
      - description: Rows per page; omit for the whole list
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.adminChallengesResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpx.ErrorBody'
      security:
      - BearerAuth: []
      summary: List challenges (admin)
      tags:
      - admin
    post:
      consumes:
      - multipart/form-data
      description: Uploads a challenge image to Cloudinary and creates/replaces the
        daily challenge for a date.
      parameters:
      - description: Challenge image (max 15MB)
        in: formData
        name: file
        required: true
        type: file
      - description: Time limit to complete, in seconds
        in: formData
        name: playSeconds
        required: true
        type: integer
      - description: Grid columns (default 6)
        in: formData
        name: gridCols
        type: integer
      - description: Grid rows (default 10)
        in: formData
        name: gridRows
        type: integer
      - description: Category (Nature/Space/Architecture/Art)
        in: formData
        name: category
        type: string
      - description: 'Challenge date YYYY-MM-DD (default: next free queue day)'
        in: formData
        name: date
        type: string
      - description: Optional title
        in: formData
        name: title
        type: string
      - description: XP awarded on completion (default 100)
        in: formData
        name: xpReward
        type: integer
      - description: Enable in-game disruption events (default false)
        in: formData
        name: challengeEventsEnabled
        type: boolean
      - description: Bonus question asked after solving (blank = no quiz)
        in: formData
        name: quizQuestion
        type: string
      - description: Answer option A (required when quizQuestion is set)
        in: formData
        name: quizOptionA
        type: string
      - description: Answer option B (required when quizQuestion is set)
        in: formData
        name: quizOptionB
        type: string
      - description: 'Which option is correct: 0 = A, 1 = B'
        in: formData
        name: quizCorrectOption
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.AdminChallengeView'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpx.ErrorBody'
      security:
      - BearerAuth: []
      summary: Create daily challenge (admin)
      tags:
      - admin
  /api/v1/admin/challenges/{id}:
    put:
      consumes:
      - multipart/form-data
      parameters:
      - description: Challenge id (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: New image (optional)
        in: formData
        name: file
        type: file
      - description: Time limit in seconds
        in: formData
        name: playSeconds
        required: true
        type: integer
      - description: Grid columns
        in: formData
        name: gridCols
        type: integer
      - description: Grid rows
        in: formData
        name: gridRows
        type: integer
      - description: Category
        in: formData
        name: category
        type: string
      - description: Title
        in: formData
        name: title
        type: string
      - description: XP awarded on completion (default 100)
        in: formData
        name: xpReward
        type: integer
      - description: Enable in-game disruption events (default false)
        in: formData
        name: challengeEventsEnabled
        type: boolean
      - description: Bonus question asked after solving (blank clears the quiz)
        in: formData
        name: quizQuestion
        type: string
      - description: Answer option A (required when quizQuestion is set)
        in: formData
        name: quizOptionA
        type: string
      - description: Answer option B (required when quizQuestion is set)
        in: formData
        name: quizOptionB
        type: string
      - description: 'Which option is correct: 0 = A, 1 = B'
        in: formData
        name: quizCorrectOption
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AdminChallengeView'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorBody'
      security:
      - BearerAuth: []
      summary: Update daily challenge (admin)
      tags:
      - admin
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the session token issued by
      /api/v1/auth/login or /register.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"